Move private function dependency to backend
